Bug #1406142 bsp: camera: fix pclint warning
[root cause  ] pclint warning
[changes     ] bsp/modules/camera
[side effects] yes
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums512
[test case]: preview
[reviewers   ] xiaobo.liu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ core/isp2.6/src/cam_core.c
@@ -2846,8 +2846,8 @@ static int camcore_cal_channel_size_rds(struct camera_module *module)
 		dcam_out.h = ALIGN(dcam_out.h, align_h);
 
 		/* keep same ratio between width and height */
-		ratio16_w = div_u64((uint64_t)trim_pv.size_x << RATIO_SHIFT, dcam_out.w);
-		ratio16_h = div_u64((uint64_t)trim_pv.size_y << RATIO_SHIFT, dcam_out.h);
+		ratio16_w = (uint32_t)div_u64((uint64_t)trim_pv.size_x << RATIO_SHIFT, dcam_out.w);
+		ratio16_h = (uint32_t)div_u64((uint64_t)trim_pv.size_y << RATIO_SHIFT, dcam_out.h);
 		ratio_min = min(ratio16_w, ratio16_h);
 
 		/* if sensor size is too small */
@@ -2898,8 +2898,8 @@ static int camcore_cal_channel_size_rds(struct camera_module *module)
 			dcam_out.h = ALIGN(dcam_out.h, align_h);
 
 			/* keep same ratio between width and height */
-			ratio16_w = div_u64((uint64_t)trim_pv.size_x << RATIO_SHIFT, dcam_out.w);
-			ratio16_h = div_u64((uint64_t)trim_pv.size_y << RATIO_SHIFT, dcam_out.h);
+			ratio16_w = (uint32_t)div_u64((uint64_t)trim_pv.size_x << RATIO_SHIFT, dcam_out.w);
+			ratio16_h = (uint32_t)div_u64((uint64_t)trim_pv.size_y << RATIO_SHIFT, dcam_out.h);
 			ratio_min = min(ratio16_w, ratio16_h);
 		}
